// libs/supabase/schema.ts
export type Database = {
// ...
  public: {
    Tables: {
      gists: {
        Row: {
          content: string
          created_at: string
          description: string
          id: string
          is_paid: boolean
          lang: string
          price: number
          profile_id: string | null
          title: string
        }
        Insert: {
          content: string
          created_at?: string
          description: string
          id: string
          is_paid: boolean
          lang: string
          price: number
          profile_id?: string | null
          title: string
        }
        Update: {
          content?: string
          created_at?: string
          description?: string
          id?: string
          is_paid?: boolean
          lang?: string
          price?: number
          profile_id?: string | null
          title?: string
        }
        Relationships: [
          {
            foreignKeyName: "gists_profile_id_fkey"
            columns: ["profile_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      sales: {
        Row: {
          created_at: string
          customer_email: string
          gist_id: string | null
          id: string
        }
        Insert: {
          created_at?: string
          customer_email: string
          gist_id?: string | null
          id: string
        }
        Update: {
          created_at?: string
          customer_email?: string
          gist_id?: string | null
          id?: string
        }
        Relationships: [
          {
            foreignKeyName: "sales_gist_id_fkey"
            columns: ["gist_id"]
            isOneToOne: false
            referencedRelation: "gists"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}